Revert "Enable debugging integration tests in VSCode"
This reverts commit 24b58b7a77b21d98d5ab82c552b28dd8678a067a.
